+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## !Deploy
+
+* Implements `-HighQuality` switch to download the high quality version of the image where available per Issue [#8](https://github.com/markwragg/Powershell-XKCD/issues/8). Thanks [@nlsdg](https://github.com/nlsdg)!
+* Fixes download of image with incorrect extension per Issue [#9](https://github.com/markwragg/Powershell-XKCD/issues/9). Thanks [@nlsdg](https://github.com/nlsdg)!
+
 ## [1.4.81] - 2020-02-25
 
 * Test deployment.